+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## (2022-03-17)
+- added azure-py-rg-storageaccounts
+
 ## (2022-03-03)
 - added azure-go-resourcegroup